requirements.txt
pandas
openpyxl
requests
Pillow